+++ src/simulation/library.jl
@@ -65,21 +65,21 @@ Wrapper containing all library construction parameters
 """
 type Library
     "Distribution of guide knockdown efficiencies"
-    knockdown_dist::Dict{Int64, Tuple{Symbol, Sampleable}}
+    knockdown_dist::Dict{Int, Tuple{Symbol, Sampleable}}
     knockdown_probs::Categorical
 
     """
     Maximum phenotype categories mapped to their probability of being
     selected and the distribution to draw from if they are selected
     """
-    max_phenotype_dists::Dict{Int64, Tuple{Symbol, Sampleable}}
+    max_phenotype_dists::Dict{Int, Tuple{Symbol, Sampleable}}
     phenotype_probs::Categorical
 
     """
     Knockdown-phenotype relationships mapped to their probability of
     being selected and their respective KDPhenotypeRelationship
     """
-    kd_phenotype_relationships::Dict{Int64, Tuple{Symbol, KDPhenotypeRelationship}}
+    kd_phenotype_relationships::Dict{Int, Tuple{Symbol, KDPhenotypeRelationship}}
     relationship_probs::Categorical
 
     """
@@ -168,7 +168,7 @@ array_names(::Type{Library}) = [:frac_hq, :mean_hq_kd, :frac_inc, :frac_dec, :ce
 
 function unroll{T}(data::Dict{Symbol, Tuple{Float64, T}})
     probs = Float64[]
-    results = Dict{Int64, Tuple{Symbol, T}}()
+    results = Dict{Int, Tuple{Symbol, T}}()
     count = 0
     for (key, value) in data
         push!(probs, value[1])
